+++ src/main/kotlin/li/auna/patches/youtube/layout/hide/subtitlestoast/fingerprints/HideSubtitlesToastFingerprint2.kt
@@ -1,0 +1,22 @@
+package li.auna.patches.youtube.layout.hide.subtitlestoast.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.MethodFingerprint
+import com.android.tools.smali.dexlib2.AccessFlags
+import com.android.tools.smali.dexlib2.Opcode
+
+object HideSubtitlesToastFingerprint2 : MethodFingerprint(
+    returnType = "V",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
+    opcodes = listOf(
+        Opcode.CONST_HIGH16,
+        Opcode.OR_INT_2ADDR,
+        Opcode.IPUT,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.CHECK_CAST
+    ),
+        customFingerprint = { methodDef, _ ->
+            methodDef.name == "onItemClick"
+        }
+)